Generalise OwnerGateway -> PersonGateway
