+++ src/main/java/uk/gov/mca/beacons/api/gateways/PersonGateway.java
@@ -4,8 +4,8 @@ import java.util.UUID;
 import uk.gov.mca.beacons.api.dto.CreateOwnerRequest;
 import uk.gov.mca.beacons.api.jpa.entities.Person;
 
-public interface OwnerGateway {
-  Person save(CreateOwnerRequest request);
+public interface PersonGateway {
+    Person save(CreateOwnerRequest request);
 
-  Person findByBeaconId(UUID beaconId);
+    Person findByBeaconId(UUID beaconId);
 }
