+++ src/main/java/uk/gov/mca/beacons/api/services/CreateOwnerService.java
@@ -3,7 +3,7 @@ package uk.gov.mca.beacons.api.services;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import uk.gov.mca.beacons.api.gateways.OwnerGateway;
+import uk.gov.mca.beacons.api.gateways.PersonGateway;
 import uk.gov.mca.beacons.api.jpa.entities.Person;
 import uk.gov.mca.beacons.api.mappers.CreateOwnerRequestMapper;
 
@@ -11,17 +11,17 @@ import uk.gov.mca.beacons.api.mappers.CreateOwnerRequestMapper;
 @Transactional
 public class CreateOwnerService {
 
-  private final OwnerGateway ownerGateway;
+    private final PersonGateway personGateway;
 
-  @Autowired
-  public CreateOwnerService(OwnerGateway ownerGateway) {
-    this.ownerGateway = ownerGateway;
-  }
+    @Autowired
+    public CreateOwnerService(PersonGateway personGateway) {
+        this.personGateway = personGateway;
+    }
 
-  public Person execute(Person person) {
-    final var createOwnerRequest = CreateOwnerRequestMapper.fromBeaconPerson(
-      person
-    );
-    return ownerGateway.save(createOwnerRequest);
-  }
+    public Person execute(Person person) {
+        final var createOwnerRequest = CreateOwnerRequestMapper.fromBeaconPerson(
+                person
+        );
+        return personGateway.save(createOwnerRequest);
+    }
 }
