+++ src/main/java/uk/gov/mca/beacons/api/services/GetBeaconsByAccountHolderIdService.java
@@ -8,7 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.mca.beacons.api.gateways.BeaconGateway;
 import uk.gov.mca.beacons.api.gateways.EmergencyContactGateway;
-import uk.gov.mca.beacons.api.gateways.OwnerGateway;
+import uk.gov.mca.beacons.api.gateways.PersonGateway;
 import uk.gov.mca.beacons.api.gateways.UseGateway;
 import uk.gov.mca.beacons.api.jpa.entities.Beacon;
 import uk.gov.mca.beacons.api.jpa.entities.BeaconUse;
@@ -17,46 +17,46 @@ import uk.gov.mca.beacons.api.jpa.entities.Person;
 @Service
 public class GetBeaconsByAccountHolderIdService {
 
-  private final BeaconGateway beaconGateway;
-  private final UseGateway useGateway;
-  private final OwnerGateway ownerGateway;
-  private final EmergencyContactGateway emergencyContactGateway;
+    private final BeaconGateway beaconGateway;
+    private final UseGateway useGateway;
+    private final PersonGateway personGateway;
+    private final EmergencyContactGateway emergencyContactGateway;
 
-  @Autowired
-  public GetBeaconsByAccountHolderIdService(
-    BeaconGateway beaconGateway,
-    UseGateway useGateway,
-    OwnerGateway ownerGateway,
-    EmergencyContactGateway emergencyContactGateway
-  ) {
-    this.beaconGateway = beaconGateway;
-    this.useGateway = useGateway;
-    this.ownerGateway = ownerGateway;
-    this.emergencyContactGateway = emergencyContactGateway;
-  }
+    @Autowired
+    public GetBeaconsByAccountHolderIdService(
+            BeaconGateway beaconGateway,
+            UseGateway useGateway,
+            PersonGateway personGateway,
+            EmergencyContactGateway emergencyContactGateway
+    ) {
+        this.beaconGateway = beaconGateway;
+        this.useGateway = useGateway;
+        this.personGateway = personGateway;
+        this.emergencyContactGateway = emergencyContactGateway;
+    }
 
-  public List<Beacon> execute(UUID accountId) {
-    final List<Beacon> beacons = beaconGateway.findAllByAccountHolderId(
-      accountId
-    );
-    if (beacons.isEmpty()) return emptyList();
-
-    beacons.forEach(
-      beacon -> {
-        final UUID beaconId = beacon.getId();
-        final List<BeaconUse> uses = useGateway.findAllByBeaconId(beaconId);
-        beacon.setUses(uses);
-
-        final Person owner = ownerGateway.findByBeaconId(beaconId);
-        beacon.setOwner(owner);
-
-        final List<Person> emergencyContacts = emergencyContactGateway.findAllByBeaconId(
-          beaconId
+    public List<Beacon> execute(UUID accountId) {
+        final List<Beacon> beacons = beaconGateway.findAllByAccountHolderId(
+                accountId
         );
-        beacon.setEmergencyContacts(emergencyContacts);
-      }
-    );
+        if (beacons.isEmpty()) return emptyList();
 
-    return beacons;
-  }
+        beacons.forEach(
+                beacon -> {
+                    final UUID beaconId = beacon.getId();
+                    final List<BeaconUse> uses = useGateway.findAllByBeaconId(beaconId);
+                    beacon.setUses(uses);
+
+                    final Person owner = personGateway.findByBeaconId(beaconId);
+                    beacon.setOwner(owner);
+
+                    final List<Person> emergencyContacts = emergencyContactGateway.findAllByBeaconId(
+                            beaconId
+                    );
+                    beacon.setEmergencyContacts(emergencyContacts);
+                }
+        );
+
+        return beacons;
+    }
 }
