+++ src/test/java/uk/gov/mca/beacons/api/services/CreateOwnerServiceUnitTest.java
@@ -10,26 +10,26 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.mca.beacons.api.dto.CreateOwnerRequest;
-import uk.gov.mca.beacons.api.gateways.OwnerGateway;
+import uk.gov.mca.beacons.api.gateways.PersonGateway;
 import uk.gov.mca.beacons.api.jpa.entities.Person;
 
 @ExtendWith(MockitoExtension.class)
 class CreateOwnerServiceUnitTest {
 
-  @Mock
-  private OwnerGateway ownerGateway;
+    @Mock
+    private PersonGateway personGateway;
 
-  @Test
-  void shouldCreateTheAccountHolder() {
-    final CreateOwnerService createOwnerService = new CreateOwnerService(
-      ownerGateway
-    );
-    final Person person = new Person();
-    final Person createdPerson = new Person();
+    @Test
+    void shouldCreateTheAccountHolder() {
+        final CreateOwnerService createOwnerService = new CreateOwnerService(
+                personGateway
+        );
+        final Person person = new Person();
+        final Person createdPerson = new Person();
 
-    given(ownerGateway.save(isA(CreateOwnerRequest.class)))
-      .willReturn(createdPerson);
+        given(personGateway.save(isA(CreateOwnerRequest.class)))
+                .willReturn(createdPerson);
 
-    assertThat(createOwnerService.execute(person), is(createdPerson));
-  }
+        assertThat(createOwnerService.execute(person), is(createdPerson));
+    }
 }
