+++ src/test/java/uk/gov/mca/beacons/api/services/GetBeaconsByAccountHolderIdServiceUnitTest.java
@@ -14,7 +14,7 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.mca.beacons.api.gateways.BeaconGateway;
 import uk.gov.mca.beacons.api.gateways.EmergencyContactGateway;
-import uk.gov.mca.beacons.api.gateways.OwnerGateway;
+import uk.gov.mca.beacons.api.gateways.PersonGateway;
 import uk.gov.mca.beacons.api.gateways.UseGateway;
 import uk.gov.mca.beacons.api.jpa.entities.Beacon;
 import uk.gov.mca.beacons.api.jpa.entities.BeaconUse;
@@ -23,96 +23,96 @@ import uk.gov.mca.beacons.api.jpa.entities.Person;
 @ExtendWith(MockitoExtension.class)
 class GetBeaconsByAccountHolderIdServiceUnitTest {
 
-  @InjectMocks
-  private GetBeaconsByAccountHolderIdService getBeaconsByAccountHolderIdService;
+    @InjectMocks
+    private GetBeaconsByAccountHolderIdService getBeaconsByAccountHolderIdService;
 
-  @Mock
-  private BeaconGateway beaconGateway;
+    @Mock
+    private BeaconGateway beaconGateway;
 
-  @Mock
-  private UseGateway useGateway;
+    @Mock
+    private UseGateway useGateway;
 
-  @Mock
-  private OwnerGateway ownerGateway;
+    @Mock
+    private PersonGateway personGateway;
 
-  @Mock
-  private EmergencyContactGateway emergencyContactGateway;
+    @Mock
+    private EmergencyContactGateway emergencyContactGateway;
 
-  @Test
-  void shouldReturnAnEmptyListIfThereAreNoBeaconsFound() {
-    final var accountId = UUID.randomUUID();
-    given(beaconGateway.findAllByAccountHolderId(accountId))
-      .willReturn(Collections.emptyList());
+    @Test
+    void shouldReturnAnEmptyListIfThereAreNoBeaconsFound() {
+        final var accountId = UUID.randomUUID();
+        given(beaconGateway.findAllByAccountHolderId(accountId))
+                .willReturn(Collections.emptyList());
 
-    final var result = getBeaconsByAccountHolderIdService.execute(accountId);
-    assertThat(result, is(empty()));
-  }
+        final var result = getBeaconsByAccountHolderIdService.execute(accountId);
+        assertThat(result, is(empty()));
+    }
 
-  @Test
-  void shouldReturnASingleBeacon() {
-    final var accountId = UUID.randomUUID();
-    final var beacon = new Beacon();
+    @Test
+    void shouldReturnASingleBeacon() {
+        final var accountId = UUID.randomUUID();
+        final var beacon = new Beacon();
 
-    given(beaconGateway.findAllByAccountHolderId(accountId))
-      .willReturn(Collections.singletonList(beacon));
+        given(beaconGateway.findAllByAccountHolderId(accountId))
+                .willReturn(Collections.singletonList(beacon));
 
-    final var result = getBeaconsByAccountHolderIdService.execute(accountId);
-    assertThat(result.size(), is(1));
-    assertThat(result.get(0), is(beacon));
-  }
+        final var result = getBeaconsByAccountHolderIdService.execute(accountId);
+        assertThat(result.size(), is(1));
+        assertThat(result.get(0), is(beacon));
+    }
 
-  @Test
-  void shouldReturnABeaconsUses() {
-    final var accountId = UUID.randomUUID();
-    final var beaconId = UUID.randomUUID();
-    final var beacon = new Beacon();
-    beacon.setId(beaconId);
-    final var beaconUse = new BeaconUse();
+    @Test
+    void shouldReturnABeaconsUses() {
+        final var accountId = UUID.randomUUID();
+        final var beaconId = UUID.randomUUID();
+        final var beacon = new Beacon();
+        beacon.setId(beaconId);
+        final var beaconUse = new BeaconUse();
 
-    given(beaconGateway.findAllByAccountHolderId(accountId))
-      .willReturn(Collections.singletonList(beacon));
-    given(useGateway.findAllByBeaconId(beaconId))
-      .willReturn(Collections.singletonList(beaconUse));
+        given(beaconGateway.findAllByAccountHolderId(accountId))
+                .willReturn(Collections.singletonList(beacon));
+        given(useGateway.findAllByBeaconId(beaconId))
+                .willReturn(Collections.singletonList(beaconUse));
 
-    final var result = getBeaconsByAccountHolderIdService.execute(accountId);
-    assertThat(result.get(0).getUses().size(), is(1));
-    assertThat(result.get(0).getUses().get(0), is(beaconUse));
-  }
+        final var result = getBeaconsByAccountHolderIdService.execute(accountId);
+        assertThat(result.get(0).getUses().size(), is(1));
+        assertThat(result.get(0).getUses().get(0), is(beaconUse));
+    }
 
-  @Test
-  void shouldReturnABeaconsOwner() {
-    final var accountId = UUID.randomUUID();
-    final var beaconId = UUID.randomUUID();
-    final var beacon = new Beacon();
-    beacon.setId(beaconId);
-    final var owner = new Person();
+    @Test
+    void shouldReturnABeaconsOwner() {
+        final var accountId = UUID.randomUUID();
+        final var beaconId = UUID.randomUUID();
+        final var beacon = new Beacon();
+        beacon.setId(beaconId);
+        final var owner = new Person();
 
-    given(beaconGateway.findAllByAccountHolderId(accountId))
-      .willReturn(Collections.singletonList(beacon));
-    given(ownerGateway.findByBeaconId(beaconId)).willReturn(owner);
+        given(beaconGateway.findAllByAccountHolderId(accountId))
+                .willReturn(Collections.singletonList(beacon));
+        given(personGateway.findByBeaconId(beaconId)).willReturn(owner);
 
-    final var result = getBeaconsByAccountHolderIdService.execute(accountId);
-    assertThat(result.get(0).getOwner(), is(owner));
-  }
+        final var result = getBeaconsByAccountHolderIdService.execute(accountId);
+        assertThat(result.get(0).getOwner(), is(owner));
+    }
 
-  @Test
-  void shouldReturnABeaconsEmergencyContacts() {
-    final var accountId = UUID.randomUUID();
-    final var beaconId = UUID.randomUUID();
-    final var beacon = new Beacon();
-    beacon.setId(beaconId);
-    final var emergencyContact = new Person();
+    @Test
+    void shouldReturnABeaconsEmergencyContacts() {
+        final var accountId = UUID.randomUUID();
+        final var beaconId = UUID.randomUUID();
+        final var beacon = new Beacon();
+        beacon.setId(beaconId);
+        final var emergencyContact = new Person();
 
-    given(beaconGateway.findAllByAccountHolderId(accountId))
-      .willReturn(Collections.singletonList(beacon));
-    given(emergencyContactGateway.findAllByBeaconId(beaconId))
-      .willReturn(Collections.singletonList(emergencyContact));
+        given(beaconGateway.findAllByAccountHolderId(accountId))
+                .willReturn(Collections.singletonList(beacon));
+        given(emergencyContactGateway.findAllByBeaconId(beaconId))
+                .willReturn(Collections.singletonList(emergencyContact));
 
-    final var result = getBeaconsByAccountHolderIdService.execute(accountId);
-    assertThat(result.get(0).getEmergencyContacts().size(), is(1));
-    assertThat(
-      result.get(0).getEmergencyContacts().get(0),
-      is(emergencyContact)
-    );
-  }
+        final var result = getBeaconsByAccountHolderIdService.execute(accountId);
+        assertThat(result.get(0).getEmergencyContacts().size(), is(1));
+        assertThat(
+                result.get(0).getEmergencyContacts().get(0),
+                is(emergencyContact)
+        );
+    }
 }
